+++ api/src/main/java/de/bsommerfeld/pathetic/api/wrapper/Depth.java
@@ -1,12 +1,9 @@
 package de.bsommerfeld.pathetic.api.wrapper;
 
-/**
- * Represents the depth of a pathfinding node or element. This class provides methods to get and
- * increment the depth value.
- */
-public class Depth {
+/** Represents the depth of a pathfinding node or element. */
+public final class Depth {
 
-  private int value;
+  private final int value;
 
   private Depth(int value) {
     this.value = value;
@@ -20,11 +17,6 @@ public class Depth {
    */
   public static Depth of(int value) {
     return new Depth(value);
-  }
-
-  /** Increments the depth value by one. */
-  public void increment() {
-    value++;
   }
 
   /**
@@ -47,23 +39,19 @@ public class Depth {
     return this.value;
   }
 
+  @Override
   public boolean equals(final Object o) {
     if (o == this) return true;
     if (!(o instanceof Depth)) return false;
     final Depth other = (Depth) o;
-    if (!other.canEqual((Object) this)) return false;
-    if (this.getValue() != other.getValue()) return false;
-    return true;
+    return this.value == other.value;
   }
 
-  protected boolean canEqual(final Object other) {
-    return other instanceof Depth;
-  }
-
+  @Override
   public int hashCode() {
     final int PRIME = 59;
     int result = 1;
-    result = result * PRIME + this.getValue();
+    result = result * PRIME + this.value;
     return result;
   }
 }